feat: fix cast long to biginteger for report context; change version to 5.1.20-SNAPSHOT;

+++ src/main/java/org/qubership/automation/itf/core/hibernate/spring/managers/reports/InstanceContextObjectManager.java
@@ -142,11 +142,11 @@ public class InstanceContextObjectManager extends AbstractObjectManager<Instance
             tcContextBriefInfo.setIniname((String) initiatorInfo.get(0)[0]);
             tcContextBriefInfo.setInitiatortype((String) initiatorInfo.get(0)[1]);
             tcContextBriefInfo.setSituationId((BigInteger) initiatorInfo.get(0)[2]);
-            tcContextBriefInfo.setChainId((BigInteger) initiatorInfo.get(0)[3]);
+            tcContextBriefInfo.setChainId(BigInteger.valueOf((Long) initiatorInfo.get(0)[3]));
             tcContextBriefInfo.setExecutiondata((String) initiatorInfo.get(0)[4]);
             tcContextBriefInfo.setOperationName((String) initiatorInfo.get(0)[5]);
             tcContextBriefInfo.setSystemName((String) initiatorInfo.get(0)[6]);
-            tcContextBriefInfo.setSystemId((BigInteger) initiatorInfo.get(0)[7]);
+            tcContextBriefInfo.setSystemId(BigInteger.valueOf((Long) initiatorInfo.get(0)[7]));
         }
         return tcContextBriefInfo;
     }
